Separate models from core to break import cycles
